Live KinectV2 skeleton data in MacOS Processing apps
Place files in a folder titled 'Kinect2UDP_receiver'. Put danceData3.csv file in folder titled 'data' within the app folder. You can run the code in Processing 2.2.1 (tested - it will crash in 3.x) and it will load and display recorded Kinect V2 data. You can set 'recorded' to false, connect a Kinect V2 to a Windows machine running the 'Kinect2UDP' application (Processing 3.x) and connect the Windows machine to the MacOS machine with ethernet (with static IP numbers set, as in the code) and you will be able to acquire live Kinect V2 skeleton data into a MacOS Processing environment (this is possible with Kinect V1 without a Windows machine but with Kinect V2 this is, as far as I know, the only solution).